+++ packages/vm/src/calls.rs
@@ -692,9 +692,10 @@ mod tests {
             call_execute::<_, _, _, Empty>(&mut instance, &mock_env(), &info, msg).unwrap_err();
         match err {
             VmError::RuntimeErr { msg, .. } => {
-                assert!(msg.contains(
-                    "RuntimeError: Aborted: panicked at 'This page intentionally faulted'"
-                ))
+                assert!(
+                    msg.contains("RuntimeError: Aborted: panicked at src/contract.rs:"),
+                    "Unexpected error msg: {msg}"
+                )
             }
             err => panic!("Unexpected error: {err:?}"),
         }
@@ -907,6 +908,7 @@ mod tests {
             );
             assert_eq!(ReplyOn::Success, res.messages[0].reply_on);
             let id = res.messages[0].id;
+            let payload = res.messages[0].payload.clone();
             let event = Event::new("instantiate").add_attributes(vec![
                 // We have to force this one to avoid the debug assertion against _
                 mock_wasmd_attr("_contract_address", account),
@@ -915,6 +917,7 @@ mod tests {
             #[allow(deprecated)]
             let response = Reply {
                 id,
+                payload,
                 gas_used: 1234567,
                 result: SubMsgResult::Ok(SubMsgResponse {
                     events: vec![event],